+++ packages/tf-inventory/bin/cli.js
@@ -15,7 +15,7 @@ program
 
 program
   .command('list-deployments')
-  .description('List Terraform deployments in the account')
+  .description('List Terraform Cumulus deployments in the account')
   .action(async () => {
     const stateFiles = await stateFile.listTfStateFiles();
     console.log(stateFile.listTfDeployments(stateFiles));
@@ -23,7 +23,7 @@ program
 
 program
   .command('deployment-report')
-  .description('List each deployment with files, number of resources, and last update date')
+  .description('List each Cumulus deployment with files, number of resources, and last update date')
   .action(async () => {
     const deployments = await stateFile.deploymentReport();
     const sortedKeys = Object.keys(deployments).sort();
@@ -35,7 +35,7 @@ program
 
 program
   .command('list-orphaned-resources')
-  .description('List resources not associated with a Terraform deployment')
+  .description('List resources not associated with a Terraform deployment, currently supports ECS and EC2')
   .action(async () => {
     console.log(await inventory.reconcileResources());
   });