+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Added `@cumulus/aws-client` package to provide utilities for working with AWS services and the Node.js AWS SDK
   - Added `@cumulus/errors` package which exports error classes for use in Cumulus workflow code
 
+- **CUMULUS-1697**
+  - Added the `@cumulus/tf-inventory` package that provides command line utilities for managing Terraform resources in your AWS account
+
 ### Changed
 
 - **CUMULUS-1686**